# src/GovUK.Dfe.FlexForms.Api.Client/Generated/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/Applications/{applicationId}/responses": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Adds a new response version to an existing application.",
        "operationId": "Applications_AddApplicationResponse",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddApplicationResponseRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "Response version created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid user token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "User does not have permission to update this application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Application not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/Applications/templates/{templateId}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "Returns a paged list of all applications for the specified template, based on the caller's role (admin or caseworker).",
        "operationId": "Applications_GetApplicationsByTemplate",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "IncludeSchema",
            "in": "query",
            "description": "When true, includes the template JSON schema in each application response.",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "PageNumber",
            "in": "query",
            "description": "One-based page number for paged results.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "Number of items per page.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "ApplicationReference",
            "in": "query",
            "description": "Partial or full application reference to search for.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "DateStartedFrom",
            "in": "query",
            "description": "Inclusive start of the date-started (created) range.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "DateStartedTo",
            "in": "query",
            "description": "Inclusive end of the date-started (created) range.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "DateSubmittedFrom",
            "in": "query",
            "description": "Inclusive start of the date-submitted range.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "DateSubmittedTo",
            "in": "query",
            "description": "Inclusive end of the date-submitted range.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "Status",
            "in": "query",
            "description": "Application status filter.",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              ]
            },
            "x-position": 10
          }
        ],
        "responses": {
          "200": {
            "description": "A paged list of applications for the template.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfApplicationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized no valid user token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/Applications/{applicationId}/file-validation-gate": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "Returns whether file validation currently blocks submit for this application.",
        "operationId": "Applications_GetFileValidationGate",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "File validation gate for the application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileValidationGateDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid user token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "403": {
            "description": "User does not have permission to read this application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Application not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceptionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "errorId": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "exceptionType": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          },
          "context": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "ApplicationResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "responseId": {
            "type": "string",
            "format": "guid"
          },
          "applicationReference": {
            "type": "string"
          },
          "applicationId": {
            "type": "string",
            "format": "guid"
          },
          "responseBody": {
            "type": "string"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "AddApplicationResponseRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "responseBody"
        ],
        "properties": {
          "responseBody": {
            "type": "string"
          }
        }
      },
...
    }
  }
}